+++ src/index.ts
@@ -1,11 +1,24 @@
-export { initializeMemoryWorkspace, type InitMemoryWorkspaceResult } from "./init/workspace.js";
-export { loadEnv } from "./config/loadEnv.js";
-export { resolveAgentDirFromEnv, resolveMemoryAgentDir, type ResolveMemoryAgentDirOptions } from "./config/agentDir.js";
-export { readPiMemoryEnv, resolveEmbedDim, type PiMemoryEnv } from "./config/env.js";
+export { initializeMemoryWorkspace, type InitMemoryWorkspaceResult } from "./store/initWorkspace.js";
+export {
+  resolveAgentDirFromEnv,
+  resolveMemoryAgentDir,
+  type ResolveMemoryAgentDirOptions,
+} from "./config/agentDir.js";
+
+export { createMemoryStore, MemoryStore } from "./store/memoryStore.js";
+
 export { runConsolidateJob, type RunConsolidateJobResult } from "./consolidate/runJob.js";
-export * from "./constants/index.js";
-export * from "./utils/index.js";
-export { MemoryStore } from "./store/memoryStore.js";
-export { ensureSidecarRunning, stopSidecar } from "./sidecar/sidecarManager.js";
-export { query, reindex, ping } from "./sidecar/client.js";
+export { mergeMemoryEntries, scheduleMergeMemoryEntriesInBackground } from "./consolidate/mergeMemoryEntries.js";
+export {
+  runDrainShutdownQueueJob,
+  type RunDrainShutdownQueueResult,
+} from "./shutdown/runDrainJob.js";
+
 export { createLlmClient, type LlmClient } from "./adapters/llm/index.js";
+
+export {
+  DEFAULT_MEMORY_FILE,
+  MEMORY_SECTIONS,
+  type MemorySection,
+} from "./constants/memory.js";
+export { PI_MEMORY_DATA_SUBDIR, PI_MEMORY_ENV_FILE_NAME } from "./constants/paths.js";